+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/api/deidServices/operations.ts
@@ -1,0 +1,363 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { HealthDataAIServicesContext as Client } from "../index.js";
+import {
+  errorResponseDeserializer,
+  DeidService,
+  deidServiceSerializer,
+  deidServiceDeserializer,
+  _DeidServiceListResult,
+  _deidServiceListResultDeserializer,
+  DeidUpdate,
+  deidUpdateSerializer,
+} from "../../models/models.js";
+import {
+  PagedAsyncIterableIterator,
+  buildPagedAsyncIterator,
+} from "../../static-helpers/pagingHelpers.js";
+import { getLongRunningPoller } from "../../static-helpers/pollingHelpers.js";
+import { expandUrlTemplate } from "../../static-helpers/urlTemplate.js";
+import {
+  DeidServicesDeleteOptionalParams,
+  DeidServicesUpdateOptionalParams,
+  DeidServicesCreateOptionalParams,
+  DeidServicesListBySubscriptionOptionalParams,
+  DeidServicesListByResourceGroupOptionalParams,
+  DeidServicesGetOptionalParams,
+} from "./options.js";
+import {
+  StreamableMethod,
+  PathUncheckedResponse,
+  createRestError,
+  operationOptionsToRequestParameters,
+} from "@azure-rest/core-client";
+import { PollerLike, OperationState } from "@azure/core-lro";
+
+export function _$deleteSend(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  options: DeidServicesDeleteOptionalParams = { requestOptions: {} },
+): StreamableMethod {
+  const path = expandUrlTemplate(
+    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.HealthDataAIServices/deidServices/{deidServiceName}{?api%2Dversion}",
+    {
+      subscriptionId: context.subscriptionId,
+      resourceGroupName: resourceGroupName,
+      deidServiceName: deidServiceName,
+      "api%2Dversion": context.apiVersion,
+    },
+    {
+      allowReserved: options?.requestOptions?.skipUrlEncoding,
+    },
+  );
+  return context.path(path).delete({ ...operationOptionsToRequestParameters(options) });
+}
+
+export async function _$deleteDeserialize(result: PathUncheckedResponse): Promise<void> {
+  const expectedStatuses = ["202", "204", "200", "201"];
+  if (!expectedStatuses.includes(result.status)) {
+    const error = createRestError(result);
+    error.details = errorResponseDeserializer(result.body);
+    throw error;
+  }
+
+  return;
+}
+
+/** Delete a DeidService */
+/**
+ *  @fixme delete is a reserved word that cannot be used as an operation name.
+ *         Please add @clientName("clientName") or @clientName("<JS-Specific-Name>", "javascript")
+ *         to the operation to override the generated name.
+ */
+export function $delete(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  options: DeidServicesDeleteOptionalParams = { requestOptions: {} },
+): PollerLike<OperationState<void>, void> {
+  return getLongRunningPoller(context, _$deleteDeserialize, ["202", "204", "200", "201"], {
+    updateIntervalInMs: options?.updateIntervalInMs,
+    abortSignal: options?.abortSignal,
+    getInitialResponse: () => _$deleteSend(context, resourceGroupName, deidServiceName, options),
+    resourceLocationConfig: "location",
+  }) as PollerLike<OperationState<void>, void>;
+}
+
+export function _updateSend(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  properties: DeidUpdate,
+  options: DeidServicesUpdateOptionalParams = { requestOptions: {} },
+): StreamableMethod {
+  const path = expandUrlTemplate(
+    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.HealthDataAIServices/deidServices/{deidServiceName}{?api%2Dversion}",
+    {
+      subscriptionId: context.subscriptionId,
+      resourceGroupName: resourceGroupName,
+      deidServiceName: deidServiceName,
+      "api%2Dversion": context.apiVersion,
+    },
+    {
+      allowReserved: options?.requestOptions?.skipUrlEncoding,
+    },
+  );
+  return context.path(path).patch({
+    ...operationOptionsToRequestParameters(options),
+    contentType: "application/json",
+    headers: {
+      accept: "application/json",
+      ...options.requestOptions?.headers,
+    },
+    body: deidUpdateSerializer(properties),
+  });
+}
+
+export async function _updateDeserialize(result: PathUncheckedResponse): Promise<DeidService> {
+  const expectedStatuses = ["200", "202", "201"];
+  if (!expectedStatuses.includes(result.status)) {
+    const error = createRestError(result);
+    error.details = errorResponseDeserializer(result.body);
+    throw error;
+  }
+
+  return deidServiceDeserializer(result.body);
+}
+
+/** Update a DeidService */
+export function update(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  properties: DeidUpdate,
+  options: DeidServicesUpdateOptionalParams = { requestOptions: {} },
+): PollerLike<OperationState<DeidService>, DeidService> {
+  return getLongRunningPoller(context, _updateDeserialize, ["200", "202", "201"], {
+    updateIntervalInMs: options?.updateIntervalInMs,
+    abortSignal: options?.abortSignal,
+    getInitialResponse: () =>
+      _updateSend(context, resourceGroupName, deidServiceName, properties, options),
+    resourceLocationConfig: "location",
+  }) as PollerLike<OperationState<DeidService>, DeidService>;
+}
+
+export function _createSend(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  resource: DeidService,
+  options: DeidServicesCreateOptionalParams = { requestOptions: {} },
+): StreamableMethod {
+  const path = expandUrlTemplate(
+    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.HealthDataAIServices/deidServices/{deidServiceName}{?api%2Dversion}",
+    {
+      subscriptionId: context.subscriptionId,
+      resourceGroupName: resourceGroupName,
+      deidServiceName: deidServiceName,
+      "api%2Dversion": context.apiVersion,
+    },
+    {
+      allowReserved: options?.requestOptions?.skipUrlEncoding,
+    },
+  );
+  return context.path(path).put({
+    ...operationOptionsToRequestParameters(options),
+    contentType: "application/json",
+    headers: {
+      accept: "application/json",
+      ...options.requestOptions?.headers,
+    },
+    body: deidServiceSerializer(resource),
+  });
+}
+
+export async function _createDeserialize(result: PathUncheckedResponse): Promise<DeidService> {
+  const expectedStatuses = ["200", "201", "202"];
+  if (!expectedStatuses.includes(result.status)) {
+    const error = createRestError(result);
+    error.details = errorResponseDeserializer(result.body);
+    throw error;
+  }
+
+  return deidServiceDeserializer(result.body);
+}
+
+/** Create a DeidService */
+export function create(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  resource: DeidService,
+  options: DeidServicesCreateOptionalParams = { requestOptions: {} },
+): PollerLike<OperationState<DeidService>, DeidService> {
+  return getLongRunningPoller(context, _createDeserialize, ["200", "201", "202"], {
+    updateIntervalInMs: options?.updateIntervalInMs,
+    abortSignal: options?.abortSignal,
+    getInitialResponse: () =>
+      _createSend(context, resourceGroupName, deidServiceName, resource, options),
+    resourceLocationConfig: "azure-async-operation",
+  }) as PollerLike<OperationState<DeidService>, DeidService>;
+}
+
+export function _listBySubscriptionSend(
+  context: Client,
+  options: DeidServicesListBySubscriptionOptionalParams = {
+    requestOptions: {},
+  },
+): StreamableMethod {
+  const path = expandUrlTemplate(
+    "/subscriptions/{subscriptionId}/providers/Microsoft.HealthDataAIServices/deidServices{?api%2Dversion}",
+    {
+      subscriptionId: context.subscriptionId,
+      "api%2Dversion": context.apiVersion,
+    },
+    {
+      allowReserved: options?.requestOptions?.skipUrlEncoding,
+    },
+  );
+  return context.path(path).get({
+    ...operationOptionsToRequestParameters(options),
+    headers: {
+      accept: "application/json",
+      ...options.requestOptions?.headers,
+    },
+  });
+}
+
+export async function _listBySubscriptionDeserialize(
+  result: PathUncheckedResponse,
+): Promise<_DeidServiceListResult> {
+  const expectedStatuses = ["200"];
+  if (!expectedStatuses.includes(result.status)) {
+    const error = createRestError(result);
+    error.details = errorResponseDeserializer(result.body);
+    throw error;
+  }
+
+  return _deidServiceListResultDeserializer(result.body);
+}
+
+/** List DeidService resources by subscription ID */
+export function listBySubscription(
+  context: Client,
+  options: DeidServicesListBySubscriptionOptionalParams = {
+    requestOptions: {},
+  },
+): PagedAsyncIterableIterator<DeidService> {
+  return buildPagedAsyncIterator(
+    context,
+    () => _listBySubscriptionSend(context, options),
+    _listBySubscriptionDeserialize,
+    ["200"],
+    { itemName: "value", nextLinkName: "nextLink" },
+  );
+}
+
+export function _listByResourceGroupSend(
+  context: Client,
+  resourceGroupName: string,
+  options: DeidServicesListByResourceGroupOptionalParams = {
+    requestOptions: {},
+  },
+): StreamableMethod {
+  const path = expandUrlTemplate(
+    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.HealthDataAIServices/deidServices{?api%2Dversion}",
+    {
+      subscriptionId: context.subscriptionId,
+      resourceGroupName: resourceGroupName,
+      "api%2Dversion": context.apiVersion,
+    },
+    {
+      allowReserved: options?.requestOptions?.skipUrlEncoding,
+    },
+  );
+  return context.path(path).get({
+    ...operationOptionsToRequestParameters(options),
+    headers: {
+      accept: "application/json",
+      ...options.requestOptions?.headers,
+    },
+  });
+}
+
+export async function _listByResourceGroupDeserialize(
+  result: PathUncheckedResponse,
+): Promise<_DeidServiceListResult> {
+  const expectedStatuses = ["200"];
+  if (!expectedStatuses.includes(result.status)) {
+    const error = createRestError(result);
+    error.details = errorResponseDeserializer(result.body);
+    throw error;
+  }
+
+  return _deidServiceListResultDeserializer(result.body);
+}
+
+/** List DeidService resources by resource group */
+export function listByResourceGroup(
+  context: Client,
+  resourceGroupName: string,
+  options: DeidServicesListByResourceGroupOptionalParams = {
+    requestOptions: {},
+  },
+): PagedAsyncIterableIterator<DeidService> {
+  return buildPagedAsyncIterator(
+    context,
+    () => _listByResourceGroupSend(context, resourceGroupName, options),
+    _listByResourceGroupDeserialize,
+    ["200"],
+    { itemName: "value", nextLinkName: "nextLink" },
+  );
+}
+
+export function _getSend(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  options: DeidServicesGetOptionalParams = { requestOptions: {} },
+): StreamableMethod {
+  const path = expandUrlTemplate(
+    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.HealthDataAIServices/deidServices/{deidServiceName}{?api%2Dversion}",
+    {
+      subscriptionId: context.subscriptionId,
+      resourceGroupName: resourceGroupName,
+      deidServiceName: deidServiceName,
+      "api%2Dversion": context.apiVersion,
+    },
+    {
+      allowReserved: options?.requestOptions?.skipUrlEncoding,
+    },
+  );
+  return context.path(path).get({
+    ...operationOptionsToRequestParameters(options),
+    headers: {
+      accept: "application/json",
+      ...options.requestOptions?.headers,
+    },
+  });
+}
+
+export async function _getDeserialize(result: PathUncheckedResponse): Promise<DeidService> {
+  const expectedStatuses = ["200"];
+  if (!expectedStatuses.includes(result.status)) {
+    const error = createRestError(result);
+    error.details = errorResponseDeserializer(result.body);
+    throw error;
+  }
+
+  return deidServiceDeserializer(result.body);
+}
+
+/** Get a DeidService */
+export async function get(
+  context: Client,
+  resourceGroupName: string,
+  deidServiceName: string,
+  options: DeidServicesGetOptionalParams = { requestOptions: {} },
+): Promise<DeidService> {
+  const result = await _getSend(context, resourceGroupName, deidServiceName, options);
+  return _getDeserialize(result);
+}