+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/api/deidServices/options.ts
@@ -1,0 +1,31 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { OperationOptions } from "@azure-rest/core-client";
+
+/** Optional parameters. */
+export interface DeidServicesDeleteOptionalParams extends OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+}
+
+/** Optional parameters. */
+export interface DeidServicesUpdateOptionalParams extends OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+}
+
+/** Optional parameters. */
+export interface DeidServicesCreateOptionalParams extends OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+}
+
+/** Optional parameters. */
+export interface DeidServicesListBySubscriptionOptionalParams extends OperationOptions {}
+
+/** Optional parameters. */
+export interface DeidServicesListByResourceGroupOptionalParams extends OperationOptions {}
+
+/** Optional parameters. */
+export interface DeidServicesGetOptionalParams extends OperationOptions {}