+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/api/healthDataAIServicesContext.ts
@@ -3,25 +3,36 @@
 
 import { logger } from "../logger.js";
 import { KnownVersions } from "../models/models.js";
+import { AzureSupportedClouds, getArmEndpoint } from "../static-helpers/cloudSettingHelpers.js";
 import { Client, ClientOptions, getClient } from "@azure-rest/core-client";
 import { TokenCredential } from "@azure/core-auth";
 
-export interface HealthDataAIServicesContext extends Client {}
+export interface HealthDataAIServicesContext extends Client {
+  /** The API version to use for this operation. */
+  /** Known values of {@link KnownVersions} that the service accepts. */
+  apiVersion: string;
+  /** The ID of the target subscription. The value must be an UUID. */
+  subscriptionId: string;
+}
 
 /** Optional parameters for the client. */
 export interface HealthDataAIServicesClientOptionalParams extends ClientOptions {
   /** The API version to use for this operation. */
   /** Known values of {@link KnownVersions} that the service accepts. */
   apiVersion?: string;
+  /** Specifies the Azure cloud environment for the client. */
+  cloudSetting?: AzureSupportedClouds;
 }
 
 export function createHealthDataAIServices(
   credential: TokenCredential,
+  subscriptionId: string,
   options: HealthDataAIServicesClientOptionalParams = {},
 ): HealthDataAIServicesContext {
-  const endpointUrl = options.endpoint ?? options.baseUrl ?? `https://management.azure.com`;
+  const endpointUrl =
+    options.endpoint ?? getArmEndpoint(options.cloudSetting) ?? "https://management.azure.com";
   const prefixFromOptions = options?.userAgentOptions?.userAgentPrefix;
-  const userAgentInfo = `azsdk-js-arm-healthdataaiservices/1.0.0-beta.1`;
+  const userAgentInfo = `azsdk-js-arm-healthdataaiservices/1.1.0-beta.1`;
   const userAgentPrefix = prefixFromOptions
     ? `${prefixFromOptions} azsdk-js-api ${userAgentInfo}`
     : `azsdk-js-api ${userAgentInfo}`;
@@ -51,5 +62,9 @@ export function createHealthDataAIServices(
       return next(req);
     },
   });
-  return clientContext;
+  return {
+    ...clientContext,
+    apiVersion,
+    subscriptionId,
+  } as HealthDataAIServicesContext;
 }